move Objects to classes.py and write Objects doc page
