+++ mkdocs.yml
@@ -116,6 +116,7 @@ nav:
         - 'Skeletonize': skeletonize.md
       - 'Naive Bayes Classifier': naive_bayes_classifier.md
       - 'Object Composition': object_composition.md
+      - 'Objects Class': objects.md
       - 'Opening': opening.md
       - 'Output Mask': output_mask.md
       - 'Outputs Class': outputs.md
