+++ plantcv/plantcv/roi/__init__.py
@@ -2,10 +2,9 @@ from plantcv.plantcv.roi.roi_methods import circle
 from plantcv.plantcv.roi.roi_methods import ellipse
 from plantcv.plantcv.roi.roi_methods import from_binary_image
 from plantcv.plantcv.roi.roi_methods import rectangle
-from plantcv.plantcv.roi.roi_methods import Objects
 from plantcv.plantcv.roi.roi_methods import auto_grid
 from plantcv.plantcv.roi.roi_methods import multi
 from plantcv.plantcv.roi.roi_methods import custom
 from plantcv.plantcv.roi.roi2mask import roi2mask
 
-__all__ = ["circle", "ellipse", "from_binary_image", "rectangle", "Objects", "auto_grid", "multi", "custom", "roi2mask"]
+__all__ = ["circle", "ellipse", "from_binary_image", "rectangle", "auto_grid", "multi", "custom", "roi2mask"]
