+++ tests/plantcv/roi/conftest.py
@@ -16,8 +16,6 @@ class RoiTestData:
         self.small_rgb_img = os.path.join(self.datadir, "setaria_small_plant_rgb.png")
         # Gray image
         self.small_gray_img = os.path.join(self.datadir, "setaria_small_plant_gray.png")
-        # Contours file
-        self.small_contours_file = os.path.join(self.datadir, "setaria_small_plant_contours.npz")
         # Grid image
         self.bin_grid_img = os.path.join(self.datadir, "brassica_2plants_bin_img.png")
